reconfigured metropolis code such that the noise distributions now return get_candidate_position() rather than a particle translation to be added to the active-particle position (as the latter does not work for sign-flip dynamics); also fixed a few small bugs

+++ src/model_settings/__init__.py
@@ -50,7 +50,7 @@ if (dimensionality_of_particle_space == 1 and type(size_of_particle_space) == fl
 elif (dimensionality_of_particle_space == 1 and type(size_of_particle_space) == int or
         (dimensionality_of_particle_space > 1 and type(size_of_particle_space) == list and
          type(size_of_particle_space[0]) == int)):
-    size_of_particle_space_over_two = np.int(0.5 * np.atleast_1d(size_of_particle_space))
+    size_of_particle_space_over_two = np.int64(0.5 * np.atleast_1d(size_of_particle_space))
 else:
     size_of_particle_space_over_two = np.atleast_1d(size_of_particle_space)
 size_of_particle_space_over_two.flags.writeable = False
